+++ r-package/R/accessibility.R
@@ -4,8 +4,9 @@
 #'              decay function. See `details` for the available decay functions.
 #'
 #' @param r5r_core a rJava object to connect with R5 routing engine
-#' @param origins,destinations a spatial sf POINT object, or a data.frame
-#'                containing the columns 'id', 'lon', 'lat'
+#' @param origins,destinations a spatial sf POINT object with WGS84 CRS, or a
+#'                             data.frame containing the columns 'id', 'lon',
+#'                             'lat'.
 #' @param opportunities_colname string. The column name in the `destinations`
 #'        input that tells the number of opportunities in each location.
 #'        Defaults to "opportunities".
@@ -95,7 +96,8 @@
 #'
 #' @details
 #'  # Decay functions:
-#'  R5 allows for multiple decay functions. More info at \url{https://docs.conveyal.com/learn-more/decay-functions}
+#'  R5 allows for multiple decay functions. More info in the original R5
+#'  documentation from Conveyal, at \url{https://docs.conveyal.com/learn-more/decay-functions}
 #'  The options include:
 #'
 #'  ## Step `step` (cumulative opportunities)
@@ -158,8 +160,9 @@
 #' or 3.
 #'
 #' The default methodology for assigning LTS values to network edges is based on
-#' commonly tagged attributes of OSM ways. See more info about LTS at
-#' \url{https://docs.conveyal.com/learn-more/traffic-stress}. In summary:
+#' commonly tagged attributes of OSM ways. See more info about LTS in the original
+#' documentation of R5 from Conveyal at \url{https://docs.conveyal.com/learn-more/traffic-stress}.
+#' In summary:
 #'
 #'- **LTS 1**: Tolerable for children. This includes low-speed, low-volume streets,
 #'  as well as those with separated bicycle facilities (such as parking-protected
@@ -206,7 +209,7 @@
 #'
 #' # build transport network
 #' data_path <- system.file("extdata/poa", package = "r5r")
-#' r5r_core <- setup_r5(data_path = data_path)
+#' r5r_core <- setup_r5(data_path = data_path, temp_dir = TRUE)
 #'
 #' # load origin/destination points
 #' points <- read.csv(file.path(data_path, "poa_hexgrid.csv"))