+++ r-package/R/transit_network_to_sf.R
@@ -22,7 +22,7 @@
 #'
 #' # build transport network
 #' path <- system.file("extdata/poa", package = "r5r")
-#' r5r_core <- setup_r5(data_path = path)
+#' r5r_core <- setup_r5(data_path = path, temp_dir = TRUE)
 #'
 #' # extract transit network from r5r_core
 #' transit_net <- transit_network_to_sf(r5r_core)