+++ r-package/tests/testthat/test-download_r5.R
@@ -6,7 +6,10 @@ testthat::skip_on_cran()
 # expected behavior
 test_that("download_r5 - expected behavior", {
 
-  testthat::expect_vector(download_r5(force_update = T))
+  testthat::expect_vector(download_r5(force_update = TRUE, temp_dir = TRUE))
+  testthat::expect_vector(download_r5(force_update = TRUE, temp_dir = FALSE))
+  testthat::expect_vector(download_r5(force_update = FALSE, temp_dir = TRUE))
+  testthat::expect_vector(download_r5(force_update = FALSE, quiet = TRUE))
 
   # testthat::expect_vector(download_r5())
   # file.remove(file.path(.libPaths()[1], "r5r", "jar", "r5r_v4.9.0.jar"))
@@ -25,5 +28,3 @@ test_that("download_r5 - expected errors", {
   testthat::expect_error(download_r5(temp_dir = 'a'))
 
   })
-
-