+++ r-package/tests/testthat/test-find_snap.R
@@ -6,7 +6,7 @@ testthat::skip_on_cran()
 # load required data and setup r5r_core
 
 data_path <- system.file("extdata/spo", package = "r5r")
-r5r_core <- setup_r5(data_path, verbose = FALSE)
+r5r_core <- setup_r5(data_path, verbose = FALSE, temp_dir = TRUE)
 points <- read.csv(file.path(data_path, "spo_hexgrid.csv"))