+++ r-package/tests_rafa/test_rafa.R
@@ -349,8 +349,9 @@ covr::function_coverage(fun=r5r::set_speed, test_file("tests/testthat/test-utils
 
 # the whole package
 Sys.setenv(NOT_CRAN = "true")
-r5r_cov <- covr::package_coverage(path = ".", type = "tests")
+r5r_cov2 <- covr::package_coverage(path = ".", type = "tests")
 r5r_cov
+
 
 x <- as.data.frame(r5r_cov)
 covr::codecov( coverage = r5r_cov, token ='2a7013e9-6562-4011-beb9-168e922c4c84' )